+++ core/block/source/bundledobjecttype.go
@@ -80,7 +80,7 @@ func (v *bundledObjectType) ReadDoc(ctx context.Context, receiver ChangeReceiver
 		s.AddRelationLinks(&model.RelationLink{Format: r.Format, Key: r.Key})
 	}
 	s.SetDetails(d)
-	s.SetDetailAndBundledRelation(bundle.RelationKeyOrigin, pbtypes.Float64(float64(model.ObjectOrigin_builtin)))
+	s.SetDetailAndBundledRelation(bundle.RelationKeyOrigin, pbtypes.Int64(int64(model.ObjectOrigin_builtin)))
 	s.SetObjectTypeKey(bundle.TypeKeyObjectType)
 	return s, nil
 }
